feat(chat): extract chat auto-scroll hook and refactor bodyMove auto-scroll logic from ChatBody into a dedicated hookuseChatAutoScroll. The new hook centralizes scroll refs, state(isAtBottom, showGoToRecent), and handlers (scrollToBottom,
updateAutoScrollState, handleWheel), plus ResizeObserver andgeneration-aware behavior. Replace the in-component implementationwith a simple import and usage to keep the body component focusedon rendering and reduce duplication.

This clarifies responsibilities, improves testability, and makesthe scrolling behavior reusable across chat UI variants.
